pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "it-helpdesk-ticket-routing-openenv"
version = "0.1.0"
description = "IT helpdesk ticket routing environment for the OpenEnv framework"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Hackstreet Boys" },
    { name = "Roopal Guha Neogi" },
    { name = "Suyash Kumar" },
]
dependencies = [
    "openenv-core @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "fastapi>=0.115",
    "pydantic>=2.7",
    "uvicorn>=0.30",
    "openai>=1.68",
    "httpx>=0.25",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = ["pytest", "httpx"]

[tool.setuptools]
py-modules = ["models", "client", "vocabulary"]

[tool.setuptools.packages.find]
include = ["server*"]
